+++ changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Donate API: View all, create and delete
 - Department API: View all, create and update
 - Event host API: View all, create and delete
+- Joined in API: View all, create and delete
 
 ### Changed
 - Path and functions in financial_transaction.php